modularized some logic & added localExceptionCounty values from CE to search result cards
