+++ backend/src/domain/training/findTrainingsBy.test.ts
@@ -6,9 +6,9 @@ import { credentialEngineUtils } from "../../credentialengine/CredentialEngineUt
 import { StubDataClient } from '../test-objects/StubDataClient';
 import ceRecords from '../test-objects/ceTestData.json'
 import expectedResult from './findTrainigsByExpectedTestData.json'
-import { zipToCounty } from './findTrainingsBy';
 
 import zipcodeJson from "../utils/zip-county.json";
+import {convertZipCodeToCounty} from "../utils/convertZipCodeToCounty";
 
 jest.mock("../../credentialengine/CredentialEngineAPI");
 
@@ -41,20 +41,5 @@ describe.skip('findTrainingsByFactory', () => {
     const findTrainingsBy = findTrainingsByFactory(dataClient);
     const trainings = await findTrainingsBy(1, ['a', 'b']);
     expect(trainings).toEqual(expectedResult);
-  });
-});
-
-describe('zipToCounty', () => {
-  it('should return the correct county for a zip code', () => {
-    const keys = Object.keys(zipcodeJson.byZip);
-    let randomZip = keys[Math.floor(Math.random() * keys.length)];
-    let county = zipcodeJson.byZip[randomZip as keyof typeof zipcodeJson.byZip]
-    expect(zipToCounty(randomZip)).toEqual(county);
-    randomZip = keys[Math.floor(Math.random() * keys.length)];
-    county = zipcodeJson.byZip[randomZip as keyof typeof zipcodeJson.byZip]
-    expect(zipToCounty(randomZip)).toEqual(county);
-    randomZip = keys[Math.floor(Math.random() * keys.length)];
-    county = zipcodeJson.byZip[randomZip as keyof typeof zipcodeJson.byZip]
-    expect(zipToCounty(randomZip)).toEqual(county);
   });
 });
