+++ backend/src/domain/training/findTrainingsBy.ts
@@ -1,10 +1,10 @@
-import { convertToTitleCaseIfUppercase } from "../utils/convertToTitleCaseIfUppercase";
+import { convertZipCodeToCounty } from "../utils/convertZipCodeToCounty";
 import { FindTrainingsBy } from "../types";
 import { Address, Training } from "./Training";
 import { CalendarLength } from "../CalendarLength";
-import { LocalException } from "./Program";
 import { DataClient } from "../DataClient";
 import { Selector } from "./Selector";
+import { getLocalExceptionCounties } from "../utils/getLocalExceptionCounties";
 import { credentialEngineAPI } from "../../credentialengine/CredentialEngineAPI";
 import { credentialEngineUtils } from "../../credentialengine/CredentialEngineUtils";
 import {
@@ -12,9 +12,6 @@ import {
   CetermsScheduleTimingType,
   CTDLResource,
 } from "../credentialengine/CredentialEngine";
-
-import zipcodeJson from "../utils/zip-county.json";
-
 
 export function getAvailableAtAddress(certificate: CTDLResource): Address {
   const availableAt = certificate["ceterms:availableAt"]?.[0];
@@ -115,12 +112,6 @@ export const findTrainingsByFactory = (dataClient: DataClient): FindTrainingsBy 
           console.log(JSON.stringify(scheduleTimingType, null, 2));
         }
 
-        const localExceptionCounties = (await dataClient.getLocalExceptionsByCip())
-          .filter((localException: LocalException) => localException.cipcode === cipCode)
-          .map((localException: LocalException) =>
-            convertToTitleCaseIfUppercase(localException.county),
-          );
-
         const training = {
           id: certificate["ceterms:ctid"],
           name: certificate["ceterms:name"] ? certificate["ceterms:name"]["en-US"] : "",
@@ -130,7 +121,7 @@ export const findTrainingsByFactory = (dataClient: DataClient): FindTrainingsBy 
             name: ownedByRecord["ceterms:name"]["en-US"],
             url: ownedByRecord["ceterms:subjectWebpage"],
             email: ownedByRecord["ceterms:email"] ? ownedByRecord["ceterms:email"][0] : null,
-            county: zipToCounty(address.zipCode),
+            county: convertZipCodeToCounty(address.zipCode),
           },
           availableAt: address,
           description: certificate["ceterms:description"]
@@ -142,7 +133,7 @@ export const findTrainingsByFactory = (dataClient: DataClient): FindTrainingsBy 
           calendarLength: await getCalendarLengthId(certificate),
           occupations: await credentialEngineUtils.extractOccupations(certificate),
           inDemand: inDemandCIPCodes.includes(cipCode ?? ""),
-          localExceptionCounty: localExceptionCounties,
+          localExceptionCounty: await getLocalExceptionCounties(dataClient, cipCode),
           tuitionCost: await credentialEngineUtils.extractCost(certificate, "costType:Tuition"),
           feesCost: await credentialEngineUtils.extractCost(certificate, "costType:MixedFees"),
           booksMaterialsCost: await credentialEngineUtils.extractCost(certificate, "costType:LearningResource"),
@@ -165,22 +156,6 @@ export const findTrainingsByFactory = (dataClient: DataClient): FindTrainingsBy 
       }),
     );
   };
-};
-
-export const zipToCounty = (zip: string | undefined): string => {
-  if (!zip) {
-    console.log("no zip found")
-    return "";
-  }
-
-  const county = zipcodeJson.byZip[zip as keyof typeof zipcodeJson.byZip];
-
-  if (!county) {
-    console.log("no county found")
-    return "";
-  }
-
-  return county;
 };
 
 /*
